Don't screw with cmemk
We're not currently using it and we have a growing suspicion
that it may be a factor in our 'stochastic failure to boot'
in that it may causes early programs (like systemd components)
to segv depending on where they land in memory?  Call it a
superstition on its way to a suspicion.