Kick display SPI up to 36MHz
Note that it looks like we could go
significantly faster than that -- up
to 48MHz -- except that, apparently,
the display's configuration gets slightly
messed up.  As if the gamma or some
other value set during init is getting
misread.

But the actually redisplaying at 48MHz
looks solid.  So, we suspect we could
leave the device tree speed at like 36MHz,
but use an ioctl to kick redisplay
xfers up to 48MHz later.  But all that
remains to be tried.